Add grayscale drawing functionality
This commit adds the ability to draw using grayscale shading mode. The getGrayscale function has been implemented to calculate a darker shade of gray based on the existing color of the grid boxes. When the shading mode is selected, the draw function now applies the grayscale effect to the grid boxes.